# Conflicts:
#	ontology/generic-function/veda-function-create.ttl
#	public/js/browser/backend.js
#	public/js/browser/individual_presenter.js
#	public/js/browser/update_service.js
#	public/js/browser/util.js

+++ control-stop.sh
@@ -5,32 +5,50 @@ TIMESTAMP=`date +%Y-%m-%d_%H_%M`
 mkdir ./logs/$TIMESTAMP
 cp ./logs/*-stderr.log ./logs/$TIMESTAMP
 
+#tarantoolctl stop init_tarantool.lua
+#pkill tarantool
+
+start-stop-daemon -Kp $PWD/.pids/veda-pid $PWD/veda
 start-stop-daemon -Kp $PWD/veda-pid $PWD/veda
-killall -9 veda
-killall -9 veda-ccus
-killall -9 veda-fanout-email
-killall -9 veda-fanout-sql
-killall -9 veda-fanout-sql-np
-killall -9 veda-fanout-sql-lp
-killall -9 veda-ft-indexer
-killall -9 veda-ltr-scripts
-killall -9 veda-scripts-lp
-killall -9 veda-scripts-main
-killall -9 veda-mstorage
-killall -9 veda-server
-killall -9 veda-ttlreader
-killall -9 veda-webserver
-killall -9 veda-gowebserver
-killall -9 veda-ft-query
-killall -9 veda-lmdb-srv
-killall -9 veda-authorization
 
-tarantoolctl stop init_tarantool.lua
-pkill tarantool
+target=".pids/"
+let count=0
+for f in "$target"/*
+do
+    pid=`cat $target/$(basename $f)`
+    echo STOP $pid $target/$(basename $f)
+    kill -kill $pid
+    let count=count+1
+done
+echo ""
+echo "Count: $count"
 
-rm .veda-pid
-rm data/module-info/*.lock
-rm data/queue/*.lock
-rm data/uris/*.lock
+rm -f -r .pids
+rm -f data/module-info/*.lock
+rm -f data/queue/*.lock
+rm -f data/uris/*.lock
+
+if [ $1 == "all" ] ; then
+    echo STOP ALL VEDA MODULES
+
+    start-stop-daemon -Kp $PWD/veda-pid $PWD/veda
+    killall -9 veda
+    killall -9 veda-ccus
+    killall -9 veda-fanout-email
+    killall -9 veda-fanout-sql
+    killall -9 veda-fanout-sql-np
+    killall -9 veda-fanout-sql-lp
+    killall -9 veda-ft-indexer
+    killall -9 veda-scripts
+    killall -9 veda-mstorage
+    killall -9 veda-server
+    killall -9 veda-ttlreader
+    killall -9 veda-webserver
+    killall -9 veda-gowebserver
+    killall -9 veda-ft-query
+    killall -9 veda-lmdb-srv
+
+fi
+
 
 exit 0
